Migrate isHHGPPMComboMove, getEstimatedRemainWeight, getActualRemainWeight to ppms.js entity file
